refactor: Service Layer 추상화 제거
