+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -1,19 +1,54 @@
 package roomescape.reservation.service;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import roomescape.reservation.service.converter.ReservationConverter;
+import roomescape.reservation.service.dto.AvailableReservationTimeServiceRequest;
+import roomescape.reservation.service.dto.CreateReservationServiceRequest;
+import roomescape.reservation.domain.ReservationId;
 import roomescape.reservation.controller.dto.AvailableReservationTimeWebResponse;
 import roomescape.reservation.controller.dto.CreateReservationWebRequest;
 import roomescape.reservation.controller.dto.ReservationResponse;
+import roomescape.reservation.service.usecase.ReservationCommandUseCase;
+import roomescape.reservation.service.usecase.ReservationQueryUseCase;
+import roomescape.theme.domain.ThemeId;
 
 import java.time.LocalDate;
 import java.util.List;
 
-public interface ReservationService {
+@Service
+@RequiredArgsConstructor
+public class ReservationService {
 
-    List<ReservationResponse> getAll();
+    private final ReservationQueryUseCase reservationQueryUseCase;
+    private final ReservationCommandUseCase reservationCommandUseCase;
 
-    List<AvailableReservationTimeWebResponse> getAvailable(LocalDate date, Long id);
+    public List<ReservationResponse> getAll() {
+        return ReservationConverter.toDto(
+                reservationQueryUseCase.getAll());
+    }
 
-    ReservationResponse create(CreateReservationWebRequest createReservationWebRequest);
+    public List<AvailableReservationTimeWebResponse> getAvailable(final LocalDate date, final Long id) {
+        final AvailableReservationTimeServiceRequest serviceRequest = new AvailableReservationTimeServiceRequest(
+                date,
+                ThemeId.from(id));
 
-    void delete(Long id);
+        return reservationQueryUseCase.getTimesWithAvailability(serviceRequest).stream()
+                .map(ReservationConverter::toWebDto)
+                .toList();
+    }
+
+    public ReservationResponse create(final CreateReservationWebRequest createReservationWebRequest) {
+        return ReservationConverter.toDto(
+                reservationCommandUseCase.create(
+                        new CreateReservationServiceRequest(
+                                createReservationWebRequest.name(),
+                                createReservationWebRequest.date(),
+                                createReservationWebRequest.timeId(),
+                                createReservationWebRequest.themeId())));
+    }
+
+    public void delete(final Long id) {
+        reservationCommandUseCase.delete(ReservationId.from(id));
+    }
 }
