--- src/main/java/roomescape/reservation/service/ReservationServiceImpl.java
@@ -1,58 +1,0 @@
-package roomescape.reservation.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import roomescape.reservation.service.converter.ReservationConverter;
-import roomescape.reservation.service.dto.AvailableReservationTimeServiceRequest;
-import roomescape.reservation.service.dto.CreateReservationServiceRequest;
-import roomescape.reservation.service.usecase.ReservationCommandUseCase;
-import roomescape.reservation.service.usecase.ReservationQueryUseCase;
-import roomescape.reservation.domain.ReservationId;
-import roomescape.reservation.controller.dto.AvailableReservationTimeWebResponse;
-import roomescape.reservation.controller.dto.CreateReservationWebRequest;
-import roomescape.reservation.controller.dto.ReservationResponse;
-import roomescape.theme.domain.ThemeId;
-
-import java.time.LocalDate;
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class ReservationServiceImpl implements ReservationService {
-
-    private final ReservationQueryUseCase reservationQueryUseCase;
-    private final ReservationCommandUseCase reservationCommandUseCase;
-
-    @Override
-    public List<ReservationResponse> getAll() {
-        return ReservationConverter.toDto(
-                reservationQueryUseCase.getAll());
-    }
-
-    @Override
-    public List<AvailableReservationTimeWebResponse> getAvailable(final LocalDate date, final Long id) {
-        final AvailableReservationTimeServiceRequest serviceRequest = new AvailableReservationTimeServiceRequest(
-                date,
-                ThemeId.from(id));
-
-        return reservationQueryUseCase.getTimesWithAvailability(serviceRequest).stream()
-                .map(ReservationConverter::toWebDto)
-                .toList();
-    }
-
-    @Override
-    public ReservationResponse create(final CreateReservationWebRequest createReservationWebRequest) {
-        return ReservationConverter.toDto(
-                reservationCommandUseCase.create(
-                        new CreateReservationServiceRequest(
-                                createReservationWebRequest.name(),
-                                createReservationWebRequest.date(),
-                                createReservationWebRequest.timeId(),
-                                createReservationWebRequest.themeId())));
-    }
-
-    @Override
-    public void delete(final Long id) {
-        reservationCommandUseCase.delete(ReservationId.from(id));
-    }
-}
