+++ src/main/java/roomescape/theme/controller/ThemeController.java
@@ -10,9 +10,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.common.utils.UriFactory;
-import roomescape.theme.service.ThemeService;
 import roomescape.theme.controller.dto.CreateThemeWebRequest;
 import roomescape.theme.controller.dto.ThemeResponse;
+import roomescape.theme.service.ThemeService;
 
 import java.net.URI;
 import java.util.List;
