+++ src/main/java/roomescape/theme/service/ThemeService.java
@@ -1,17 +1,47 @@
 package roomescape.theme.service;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import roomescape.reservation.domain.ReservationDate;
+import roomescape.theme.service.converter.ThemeConverter;
+import roomescape.theme.domain.ThemeId;
 import roomescape.theme.controller.dto.CreateThemeWebRequest;
 import roomescape.theme.controller.dto.ThemeResponse;
+import roomescape.theme.service.usecase.ThemeCommandUseCase;
+import roomescape.theme.service.usecase.ThemeQueryUseCase;
 
+import java.time.LocalDate;
 import java.util.List;
 
-public interface ThemeService {
+@Service
+@RequiredArgsConstructor
+public class ThemeService {
 
-    List<ThemeResponse> getAll();
+    private final ThemeQueryUseCase themeQueryUseCase;
+    private final ThemeCommandUseCase themeCommandUseCase;
 
-    List<ThemeResponse> getRanking();
+    public List<ThemeResponse> getAll() {
+        return ThemeConverter.toDto(themeQueryUseCase.getAll());
+    }
 
-    ThemeResponse create(CreateThemeWebRequest createThemeWebRequest);
+    public List<ThemeResponse> getRanking() {
+        final int count = 10;
+        final int dateRange = 7;
+        final LocalDate endDate = LocalDate.now();
+        final LocalDate startDate = endDate.minusDays(dateRange);
 
-    void delete(Long id);
+        return ThemeConverter.toDto(
+                themeQueryUseCase.getRanking(
+                        ReservationDate.from(startDate),
+                        ReservationDate.from(endDate),
+                        count));
+    }
+
+    public ThemeResponse create(final CreateThemeWebRequest createThemeWebRequest) {
+        return ThemeConverter.toDto(themeCommandUseCase.create(ThemeConverter.toServiceDto(createThemeWebRequest)));
+    }
+
+    public void delete(final Long id) {
+        themeCommandUseCase.delete(ThemeId.from(id));
+    }
 }
