+++ src/main/java/roomescape/theme/service/usecase/ThemeCommandUseCase.java
@@ -1,12 +1,24 @@
 package roomescape.theme.service.usecase;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import roomescape.theme.service.converter.ThemeConverter;
 import roomescape.theme.service.dto.CreateThemeServiceRequest;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.ThemeId;
+import roomescape.theme.repository.ThemeRepository;
 
-public interface ThemeCommandUseCase {
+@RequiredArgsConstructor
+@Service
+public class ThemeCommandUseCase {
 
-    Theme create(CreateThemeServiceRequest createThemeServiceRequest);
+    private final ThemeRepository themeRepository;
 
-    void delete(ThemeId id);
+    public Theme create(final CreateThemeServiceRequest createThemeServiceRequest) {
+        return themeRepository.save(ThemeConverter.toDomain(createThemeServiceRequest));
+    }
+
+    public void delete(final ThemeId id) {
+        themeRepository.deleteById(id);
+    }
 }
