+++ src/main/java/roomescape/theme/service/usecase/ThemeQueryUseCase.java
@@ -1,16 +1,36 @@
 package roomescape.theme.service.usecase;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import roomescape.reservation.service.dto.ThemeToBookCountServiceResponse;
 import roomescape.reservation.domain.ReservationDate;
+import roomescape.reservation.service.usecase.ReservationQueryUseCase;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.ThemeId;
+import roomescape.theme.repository.ThemeRepository;
 
 import java.util.List;
 
-public interface ThemeQueryUseCase {
+@Service
+@RequiredArgsConstructor
+public class ThemeQueryUseCase {
 
-    List<Theme> getAll();
+    private final ThemeRepository themeRepository;
+    private final ReservationQueryUseCase reservationQueryUseCase;
 
-    Theme get(ThemeId id);
+    public List<Theme> getAll() {
+        return themeRepository.findAll();
+    }
 
-    List<Theme> getRanking(ReservationDate startDate, ReservationDate endDate, int count);
+    public Theme get(final ThemeId id) {
+        return themeRepository.findById(id)
+                .orElseThrow();
+    }
+
+    public List<Theme> getRanking(final ReservationDate startDate, final ReservationDate endDate, final int count) {
+        return reservationQueryUseCase.getRanking(startDate, endDate, count).stream()
+                .map(ThemeToBookCountServiceResponse::theme)
+                .toList();
+    }
+
 }
