+++ src/main/java/roomescape/time/service/ReservationTimeService.java
@@ -1,15 +1,37 @@
 package roomescape.time.service;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import roomescape.time.service.converter.ReservationTimeConverter;
+import roomescape.time.service.dto.CreateReservationTimeServiceRequest;
+import roomescape.time.domain.ReservationTimeId;
 import roomescape.time.controller.dto.CreateReservationTimeWebRequest;
 import roomescape.time.controller.dto.ReservationTimeResponse;
+import roomescape.time.service.usecase.ReservationTimeCommandUseCase;
+import roomescape.time.service.usecase.ReservationTimeQueryUseCase;
 
 import java.util.List;
 
-public interface ReservationTimeService {
+@Service
+@RequiredArgsConstructor
+public class ReservationTimeService {
 
-    List<ReservationTimeResponse> getAll();
+    private final ReservationTimeQueryUseCase reservationTimeQueryUseCase;
+    private final ReservationTimeCommandUseCase reservationTimeCommandUseCase;
 
-    ReservationTimeResponse create(CreateReservationTimeWebRequest createReservationTimeWebRequest);
+    public List<ReservationTimeResponse> getAll() {
+        return ReservationTimeConverter.toDto(
+                reservationTimeQueryUseCase.getAll());
+    }
 
-    void delete(Long id);
+    public ReservationTimeResponse create(final CreateReservationTimeWebRequest createReservationTimeWebRequest) {
+        return ReservationTimeConverter.toDto(
+                reservationTimeCommandUseCase.create(
+                        new CreateReservationTimeServiceRequest(
+                                createReservationTimeWebRequest.startAt())));
+    }
+
+    public void delete(final Long id) {
+        reservationTimeCommandUseCase.delete(ReservationTimeId.from(id));
+    }
 }
