+++ src/test/java/roomescape/theme/service/ThemeQueryUseCaseTest.java
@@ -10,7 +10,7 @@ import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationDate;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservation.domain.ReserverName;
-import roomescape.theme.service.usecase.ThemeQueryUseCaseImpl;
+import roomescape.theme.service.usecase.ThemeQueryUseCase;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.ThemeDescription;
 import roomescape.theme.domain.ThemeName;
@@ -27,10 +27,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
 @Transactional
-class ThemeQueryUseCaseImplTest {
+class ThemeQueryUseCaseTest {
 
     @Autowired
-    private ThemeQueryUseCaseImpl themeQueryUseCase;
+    private ThemeQueryUseCase themeQueryUseCase;
 
     @Autowired
     private ThemeRepository themeRepository;
